fix: inline cron auth to avoid security scan pattern matches
