--- src/lib/cron-auth.ts
@@ -1,25 +1,0 @@
-import { NextRequest, NextResponse } from "next/server";
-import crypto from "crypto";
-
-function timingSafeEqual(a: string, b: string): boolean {
-  const bufA = Buffer.from(a);
-  const bufB = Buffer.from(b);
-  if (bufA.byteLength !== bufB.byteLength) return false;
-  return crypto.timingSafeEqual(bufA, bufB);
-}
-
-export function verifyCronAuth(request: NextRequest): NextResponse | null {
-  const secret = process.env.CRON_SECRET;
-  if (!secret) {
-    return NextResponse.json({ error: "Server misconfigured" }, { status: 500 });
-  }
-
-  const auth = request.headers.get("authorization") ?? "";
-  const expected = `Bearer ${secret}`;
-
-  if (auth.length !== expected.length || !timingSafeEqual(auth, expected)) {
-    return NextResponse.json({ error: "Unauthorized" }, { status: 401 });
-  }
-
-  return null;
-}
